third_party/blink/web_tests/external/wpt/html/canvas/element/text-styles/2d.text.draw.baseline.ideographic-expected.txt

This is a testharness.js-based test.
FAIL Canvas test: 2d.text.draw.baseline.ideographic assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
